backend/pyproject.toml
[project]
name = "sentio-backend"
version = "0.1.0"
description = "Backend for Sentio using LangChain and FastAPI"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-chroma",
    "chromadb",
    "langchain-text-splitters",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.9",
    "langgraph>=0.2.0",
    "mcp>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "tavily-python>=0.3.0",
    "pyyaml>=6.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
